aml_django_simulator/requirements.txt
annotated-types==0.7.0
asgiref==3.11.0
cffi==2.0.0
cryptography==46.0.3
Django==6.0.1
django-cors-headers==4.9.0
django-environ==0.12.0
django-ninja==1.5.3
oracledb==3.4.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
gunicorn==21.2.0
whitenoise==6.6.0
